requirements.txt
email
imaplib
python-dotenv